let code work even if redis container is not available

+++ app/redis_db/utils.py
@@ -22,26 +22,38 @@ def redis_connect() -> redis.client.Redis:
             return client
         raise Exception(redis.RedisError)
     except redis.AuthenticationError as redis_auth_error:
-        log.error("AuthenticationError xxxxxxxxxxxxxxx")
+        log.error("Auth error from Redis!!!")
         raise UnAuthorizedException("Redis Connection Failed") from redis_auth_error
+    except Exception as any_error:
+        log.error("Redis connection failed. May be Redis container is not running at redis:6379 \
+correct details below.")
+        log.error(any_error)
+        # not raisig error to be able to function even without redis, eg. local dev
+        return None 
 
 def get_routes_from_cache(key: str):
     """Data from redis."""
     log.info('In redis get data')
     redis_client = redis_connect()
-    val = redis_client.get(key)
+    val = None
+    if redis_client is not None:
+        val = redis_client.get(key)
     return val
 
 def set_routes_to_cache(key: str, value: str):
     """Data to redis."""
     log.info('In redis set data')
     redis_client = redis_connect()
-    state = redis_client.setex(key, timedelta(seconds=180), value=value)
+    state = None
+    if redis_client is not None:
+        state = redis_client.setex(key, timedelta(seconds=180), value=value)
     return state
 
 def del_cache(key: str):
     """del cache data"""
     log.info('In redis delete data')
     redis_client = redis_connect()
-    val = redis_client.delete(key)
+    val = None
+    if redis_client is not None:
+        val = redis_client.delete(key)
     return val
